create activity document
