+++ packages/app/src/server/service/in-app-notification.ts
@@ -68,6 +68,11 @@ export default class InAppNotificationService {
     return;
   }
 
+  // inAppNotificationSchema.virtual('actionUsers').get(function(this: InAppNotificationDocument) {
+  //   const Activity = getModelSafely('Activity') || require('../models/activity')(this.crowi);
+  //   return Activity.getActionUsersFromActivities((this.activities as any) as ActivityDocument[]);
+  // });
+
 }
 
 module.exports = InAppNotificationService;
